Use reenterant versions of time functions!
Replace use of gmtime, localtime and asctime in sconex/Time and sconex/Date
with the equivalent _r versions. This was causing the incorrect time to be
returned infrequently and randomly, probably manifesting in a range of bugs.
One in particular, which resulting in me finding this issue, was that
frequently used HTTP sessions would randomly expire after a while.