New asynchronous http::Session locking in sconesite
Remove code which spins waiting to lock sessions, as this can cause the whole
server to lockup if all threads get into this state. Sessions are now only
locked when required (i.e. before processing in sconesite), and this is done
asynchronously, returning from evemt processing if the session cannot be locked
and retrying later.