Add sconectl - a config/control client for sconeserver
sconectl is a simple perl script which sends a command to
sconeserver via a local config socket, and prints the result.
A suitable local socket (/var/run/sconectl.socket) is now created
by default in sconeserver.conf. The socket is create with 0700
permissions, so sconectl should only be usable by the superuser.