Refactor sconesite ArticleBody
ArticleBody is renamed Document, and a lot of the file management
has been moved here, allowing XMLArticleBody to be removed and
XMLDoc to be simplified.
WikiTextArticleBody has been removed along with support for wtx
articles while this is being reworked as a Document class.
ArticleHeading is renamed Heading for clarity.