configure: fix static linking with openSSL
When doing static builds, the ordering of libs is important, as the
linker does not back-pedal to previous libraries specified on the
command line to find missing symbols, and only searches for those
missing symbols in the following libs.

Thus, as -lssl needs symbols from -lcrypto, it needs to come before
-lcrypto. Also, -lcrypto needs -lz, which is entirely missing.

Instead of trying to hand-craft openSSL detection, use pkg-config to
find it; pkg-config will give us all the libraries that are needed to
link in -lssl , without us double-guessing anything! :-)

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>